[receiver/statsdreceiver] Gracefully handle empty tagset (#27011)
We have a set of client libraries
(`com.timgroup.statsd.NonBlockingStatsDClient`) that send a metric like
this where there are no tags:

```
some_metric_name:1|c|#
```

Now the validity of this StatsD message can be questioned but the fact
of the matter is this is widespread/common behavior for now. This PR
includes a fix to move on if no tags in the array are present.