[chore][pkg/stanza] Extract max length trimming responsibility into trim package (#27260)
This extracts the "max length" responsibility from the scanner. A couple
benefits of this:
1. The function can be applied _after_ the flush function. Previously,
if a token was flushed due to timing, we were not enforcing the max
length.
2. It makes the function directly testable.
4. We can simplify the scanner constructor, which previously had a
parameter for the buffer size which was only used for testing the
behavior of the extracted function. This is now tested within the trim
function.