[chore][pkg/stanza] Ensure all start/stop calls in are balanced (#28294)
Follows
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/28228

This normalizes calls to `Start` and `Stop` across the test suite. 

In some cases, `poll` is called directly in order to trigger behavior
independently of timing. However, we should _either_ use `poll`
directly, or use both `Start` and `Stop` exactly once. In the future, I
expect `poll` will be exported and tested directly as part of an
internal package.