[chore] bump go versions in workflows to 1.20.11 and 1.21.4 (#29080)
This fixes security vulnerabilities found via govulncheck in the
standard library when running against the previous patch versions of
golang. While these vulnerabilities don't actually present themselves in
the binary, the workflows when running govuln check fail and thus taking
in the latest patches fix the issue.


Testing gets caught in workflow run. Noticed the issue originally when
running workflows on this pr:
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/28885